Add SpeedConfig property to CPU operating config interface
Fixes nvbug https://nvbugs/3651833 and https://nvbugs/3675974

This commit adds SpeedConfig property which combines SpeedLocked and SpeedLimit
values as a tuple.
It helps to set the speed lock and speed limit in a single dbus set property.

TODO: Reuse existing speed lock and speed limit property and add the logic
in bmcweb to decide the speed limit has to be locked or not.

Signed-off-by: Aishwary Joshi <aishwaryj@nvidia.com>