PDC Create dump interface can throw Unavailable error when another dump is in progress. This would be translated resource in use bmcweb error
tested
trigger RF bmc dump when another one is in progress.
we get below error when dump is in progress.

{
"error": {
"@Message.ExtendedInfo": [
{
"@odata.type": "#Message.v1_1_1.Message",
"Message": "The change to the requested resource failed because the resource is in use or in transition.",
"MessageArgs": [],
"MessageId": "Base.1.13.0.ResourceInUse",
"MessageSeverity": "Warning",
"Resolution": "Remove the condition and resubmit the request if the operation failed."
}
],
"code": "Base.1.13.0.ResourceInUse",
"message": "The change to the requested resource failed because the resource is in use or in transition."
}
}

Fixes nvbug https://nvbugswb.nvidia.com/NvBugs5/SWBug.aspx?bugid=4019328&cmtNo=
Signed-off-by: Rohit PAI <ropai@nvidia.com>