Update MemoryChannelPresence yaml
MemoryChannelPresence is set to boolean currently because of which it is
set to false when the presence information is not available. This gives
a false impression that the it is not present.

Changed the values to enum and added Unavailable.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5225477

Signed-off-by: Karthick Sundarrajan <ksundarrajan@nvidia.com>


(cherry picked from commit f7e1df029d5f5e20bcf2599d5386c30d31260e04)

Co-authored-by: Karthick Sundarrajan <ksundarrajan@nvidia.com>