Bluefield Rshim interface
Add Rshim interface for bluefield Rshim related functions, this
interface contains 'State' and 'LfwpEnabled' property.

For 'State' property, it is used for enable/disable/force-enable
Rshim on BMC. A enumeration 'State' parameter is defined for the
property to distinguish the action.

The 'LfwpEnabled' property is to control and indicate the LFWP status.

Fixes jira https://redmine.mellanox.com/issues/4486136

Signed-off-by: Sean Zhang <xiazhang@nvidia.com>