[meta-gb200-nvl] Update NVLink LED status Interface
Update NVLink LED status Interface

Request:
There are NVLink status LED's on the front of the chassis. For
remote monitoring it has been requested to make these visible
via the redfish interface.

Solution:
The chassis LED's are controlled by the FPGA. The LED state and
inputs used to generate that state can be read via NSM. We
read this state and expose it via dbus. Bmcweb then needs to
read this interface and show the values in chassis.

Changes:
Updated the Enum to match the LED states in the spec

Tested:
Build, loaded on system

Fixes Jira https://jirasw.nvidia.com/browse/DGXOPENBMC-12593

Signed-off-by:  dmcclung@nvidia.com